fix: fix get user email logic and octokit log
